+++ crates/interpreter/src/format.rs
@@ -1,4 +1,4 @@
-// Copyright 2023 Google LLC
+// Copyright 2025 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,13 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use rand::RngCore;
-use wasefire_applet_api::rng as api;
+use alloc::vec::Vec;
 
-#[unsafe(no_mangle)]
-unsafe extern "C" fn env_rb(params: api::fill_bytes::Params) -> isize {
-    let api::fill_bytes::Params { ptr, len } = params;
-    let buf = unsafe { std::slice::from_raw_parts_mut(ptr, len) };
-    rand::rng().fill_bytes(buf);
-    0
+pub fn leb128(wasm: &mut Vec<u8>, mut x: usize) {
+    assert!(x <= u32::MAX as usize);
+    while x > 127 {
+        wasm.push(0x80 | (x & 0x7f) as u8);
+        x >>= 7;
+    }
+    wasm.push(x as u8);
+}
+
+pub fn custom_section(wasm: &mut Vec<u8>, name: &str, content: &[u8]) {
+    assert!(name.len() < 128);
+    wasm.push(0);
+    leb128(wasm, 1 + name.len() + content.len());
+    wasm.push(name.len() as u8);
+    wasm.extend_from_slice(name.as_bytes());
+    wasm.extend_from_slice(content);
 }