+++ crates/interpreter/tests/spec.rs
@@ -58,7 +58,7 @@ fn test(repo: &str, name: &str, skip: usize) {
             _ => unimplemented!("{:?}", directive),
         }
     }
-    assert_eq!(env.skip, skip);
+    assert_eq!(env.skip, skip, "actual vs expected number of unsupported (and skipped) tests");
 }
 
 fn pool_size(name: &str) -> usize {
@@ -164,8 +164,9 @@ impl<'m> Env<'m> {
     }
 
     fn maybe_instantiate(&mut self, name: &str, wasm: &[u8]) -> Result<InstId, Error> {
+        let wasm = prepare(wasm)?;
         let module = self.alloc(wasm.len());
-        module.copy_from_slice(wasm);
+        module.copy_from_slice(&wasm);
         let module = Module::new(module)?;
         let memory = self.alloc(mem_size(name));
         self.store.instantiate(module, memory)
@@ -337,13 +338,13 @@ fn assert_invoke(env: &mut Env, invoke: WastInvoke) {
 
 fn assert_malformed(env: &mut Env, mut wat: QuoteWat) {
     if let Ok(wasm) = wat.encode() {
-        assert_eq!(only_sup!(env, Module::new(&wasm)).err(), Some(Error::Invalid));
+        assert_eq!(only_sup!(env, prepare(&wasm)).err(), Some(Error::Invalid));
     }
 }
 
 fn assert_invalid(env: &mut Env, mut wat: QuoteWat) {
     let wasm = wat.encode().unwrap();
-    assert_eq!(only_sup!(env, Module::new(&wasm)).err(), Some(Error::Invalid));
+    assert_eq!(only_sup!(env, prepare(&wasm)).err(), Some(Error::Invalid));
 }
 
 fn assert_exhaustion(env: &mut Env, call: WastInvoke) {
@@ -441,7 +442,8 @@ test!(binary_leb128, "binary-leb128");
 test!(block);
 test!(br);
 test!(br_if);
-test!(br_table);
+// TODO(dev/fast-interp): Don't skip tests we want to support.
+test!(br_table; 149);
 test!(bulk);
 test!(call);
 test!(call_indirect);