+++ crates/xtask/src/main.rs
@@ -493,6 +493,8 @@ impl RunnerOptions {
             rustflags.push("-C link-arg=-Tlink.x".to_string());
             if main.release {
                 cargo.arg("-Zbuild-std=core,alloc");
+                // TODO(https://github.com/rust-lang/rust/issues/122105): Remove when fixed.
+                rustflags.push("--allow=unused-crate-dependencies".to_string());
                 let mut features = "-Zbuild-std-features=panic_immediate_abort".to_string();
                 if self.opt_level.is_some_and(action::OptLevel::optimize_for_size) {
                     features.push_str(",optimize_for_size");
@@ -663,6 +665,8 @@ impl RunnerOptions {
             cargo.current_dir("crates/runner-nordic/crates/bootloader");
             cargo.args(["build", "--release", "--target=thumbv7em-none-eabi"]);
             cargo.args(["-Zbuild-std=core", "-Zbuild-std-features=panic_immediate_abort"]);
+            // TODO(https://github.com/rust-lang/rust/issues/122105): Remove when fixed.
+            cargo.env("RUSTFLAGS", "--allow=unused-crate-dependencies");
             cmd::execute(&mut cargo).await?;
             tokio::task::spawn_blocking(move || {
                 anyhow::Ok(flashing::download_file(
@@ -759,7 +763,7 @@ async fn wrap_command() -> Result<Command> {
 }
 
 async fn ensure_assemblyscript() -> Result<()> {
-    const ASC_VERSION: &str = "0.27.31"; // scripts/upgrade.sh relies on this name
+    const ASC_VERSION: &str = "0.27.34"; // scripts/upgrade.sh relies on this name
     const BIN: &str = "examples/assemblyscript/node_modules/.bin/asc";
     const JSON: &str = "examples/assemblyscript/node_modules/assemblyscript/package.json";
     if fs::exists(BIN).await && fs::exists(JSON).await {